ai_voice_chatbot/requirements.txt
certifi==2024.7.4
Flask==3.0.3
langchain_core==0.2.21
openai==1.35.14
pyngrok==7.1.6
python-dotenv==1.0.1
Requests==2.32.3
twilio==6.0.0
